Add GPS location clustering api.
Motivation:
To display on the map the locations of the photos in the library.
The limitation of naively putting a marker on the map for every photo in the library is that if there are thousands of points, it slows down the frontend.

A reasonable solution is that we group the gps coordiates of the photos, and only use the cluster means. This way we won't have the exact positions of individual photos, but we can see roughly where the user has taken photos.