1. Use MIT's places365 pretrained model for additional "caption" information.
2. Add AlbumThing model. This works similar to AlbumDate, in the
following way:

 - On each photo added, we extract captions (densecap AND places365)
 - Use Spacy to extract lemmatized nouns (probably an overkill to use
   spacy).
 - For each lemmatized nouns in the caption, create or get the unique
   AlbumThing object, identified by the noun.
 - Add the photo to each of the AlbumThing objects that correspond to
   nouns in the captions of this photo.

The serializer and viewset are almost identical to that of AlbumDate,
except we return a list of maximum four cover photo urls, instead of one
cover photo url. The reasoning for this is the following:

When we show a list of things (nouns) on the frontend, if we only show
one photo (not the region of the photo corresponding to the thing, but
the whole photo itself), then there is a high probability that the
cover photo representing the thing actually has nothing to do with the
thing, because the object recognizers spew out a lot of "wrong" labels.