Add reasonableness check for filename parsed dates
Dates in the far distant future should not be considered reasonable and
should not be used. A date in the far future likely indicates a false
positive for the regular expression rather than time travel.

The change includes a 30 day fudge factor which should more than
accommodate reasonable clock drift on digital cameras that don't get
network time updates.